Unseperate the chained classes
For the classes, I misunderstood what the README
instructions were asking. When it said edit
elements with both said classes, I thought it meant
elements with at least ONE of the classes, but it
actually meant elements with both classes
SIMUTANEOUSLY. So, I chained the classes together
in CSS by removing the comas that seperated them.